# docs/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "信号检修封锁窗口许可服务（本地模拟演练）",
    "description" : "仅供本地模拟演练的封锁窗口许可服务，**不连接、不控制任何真实铁路设备**。\n\n业务规则要点：\n* 一个计划可占用多个相邻区段；\n* 窗口重叠不必然冲突——必须按“互斥作业矩阵”逐对判断，禁止只比设备编号；\n* 开工前须同时满足：窗口已开放、各区段保护到位、人员到岗且资质覆盖到预计结束时刻；\n* 资质在预计结束前失效的计划必须重新安排；\n* 销记前逐项确认复核与人员撤离；计划销记后迟到的开工消息一律拒绝。\n",
    "contact" : {
      "name" : "Signal Maintenance Planner (simulation)"
    },
    "license" : {
      "name" : "Simulation-only, no operational use"
    },
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
...
  "paths" : {
    "/api/reference/sections" : {
      "get" : {
        "tags" : [ "基础数据" ],
        "summary" : "区段列表",
        "operationId" : "listSections",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/SectionView"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "基础数据" ],
        "summary" : "登记区段（可指定相邻区段 adjacentTo，多区段计划会校验连通性）",
        "operationId" : "createSection",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SectionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SectionView"
                }
              }
            }
          }
        }
      }
    },
    "/api/reference/qualifications" : {
      "get" : {
        "tags" : [ "基础数据" ],
        "summary" : "资质列表",
        "operationId" : "listQualifications",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/QualificationView"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "基础数据" ],
        "summary" : "为人员登记作业资质及有效期",
        "operationId" : "addQualification",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/QualificationRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/QualificationView"
                }
              }
            }
          }
        }
      }
    },
    "/api/reference/persons" : {
      "get" : {
        "tags" : [ "基础数据" ],
        "summary" : "人员列表",
        "operationId" : "listPeople",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/PersonView"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "基础数据" ],
        "summary" : "登记人员",
        "operationId" : "createPerson",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PersonRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PersonView"
                }
              }
            }
          }
        }
      }
    },
    "/api/reference/mutex-rules" : {
      "get" : {
        "tags" : [ "基础数据" ],
        "summary" : "互斥作业矩阵",
        "operationId" : "listMutexRules",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/MutexRuleView"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "基础数据" ],
        "summary" : "登记/更新互斥作业矩阵条目（冲突判断的唯一权威依据，不允许只比设备编号）",
        "operationId" : "putMutexRule",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/MutexRuleRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MutexRuleView"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "SectionRequest" : {
        "required" : [ "code", "name" ],
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "adjacentTo" : {
            "type" : "string"
          }
        }
      },
      "SectionView" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "adjacentTo" : {
            "type" : "string"
          }
        }
      },
      "QualificationRequest" : {
        "required" : [ "personId", "validFrom", "validUntil", "workType" ],
        "type" : "object",
        "properties" : {
          "personId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "workType" : {
            "type" : "string"
          },
          "validFrom" : {
            "type" : "string",
            "format" : "date-time"
          },
          "validUntil" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "QualificationView" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "personId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "workType" : {
            "type" : "string"
          },
          "validFrom" : {
            "type" : "string",
            "format" : "date-time"
          },
          "validUntil" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "PersonRequest" : {
        "required" : [ "name", "role" ],
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "role" : {
            "type" : "string"
          }
        }
      },
      "PersonView" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "name" : {
            "type" : "string"
          },
          "role" : {
            "type" : "string"
          }
        }
      },
      "MutexRuleRequest" : {
        "required" : [ "workTypeA", "workTypeB" ],
        "type" : "object",
        "properties" : {
          "workTypeA" : {
            "type" : "string"
          },
          "workTypeB" : {
            "type" : "string"
          },
          "exclusive" : {
            "type" : "boolean"
          },
          "remark" : {
            "type" : "string"
          }
        }
      },
      "MutexRuleView" : {
        "type" : "object",
        "properties" : {
          "workTypeA" : {
            "type" : "string"
          },
          "workTypeB" : {
            "type" : "string"
          },
          "exclusive" : {
            "type" : "boolean"
          },
          "remark" : {
            "type" : "string"
          }
        }
      },
...
    }
  }
}